requirements.txt
flask-cors
fasttext-wheel
scikit-learn
pandas
flask